Introduces asynchronous logging
Where AsyncLogProvider wraps an actual LogProvider, and internally all sub
components thereof, providing the asynchronous aspect of the logging.
Each logging statement is registered as an event, to be processed by
a separate and dedicated logging thread in the background.

The driving reason for this is to avoid I/O waits on critical execution
paths that does, even ever so little, logging.

Implementation uses AsyncEvents, which is a lock-free unbounded queue with
a consumer for processing the log events, i.e. making the actual logging
happen. Logging in neo4j is considered to be scarse enough to not impose a
heap usage problem which may theoretically be a problem for an unbounded
queue.