Greatly improves insertion performance of GB+Tree
By bringing worst-case performance much closer to best-case.
A GB+Tree consists of internal and leaf nodes, each 8k in size. Keys and values
in each node are physically ordered, order dictated by the Layout instance
provided to the tree instance. This means that inserting/removing an entry
(key/value or key/child) to the far "right", i.e. in the higher order of a node
is very cheap due to only few entries right of it will need to move.
Inserting/removing an entry to the far left, i.e. in the lower order of a node
is quite costly due to the majority of the entries in the node will need to move.

In scenarios where there's less tree traversing compared to number of insertions,
i.e. not completely randomly ordered keys to insert/remove, this gap between
best and worst case is very visible and can be 50x or even higher.

This set of changes bridges this gap from 50x down to roughly 5x, where best case
takes a small hit of a couple of percent degradation from current performance.
All in all the average insertion performance of a GB+Tree increases at least
a couple of times.

The change is to introduce a DELTA section in each leaf node (not internals a.t.m.
because limiting complexity and the majority of writes happens in leaf nodes).

           ==>                                           <==
  ┌────────┬───┬───┬───┬───┬───┬───────────────────┬───┬───┐
  │ HEADER │ A │ D │ E │ K │ M │                   │ F │ B │
  └────────┴───┴───┴───┴───┴───┴───────────────────┴───┴───┘
           <------- MAIN ------>                   < DELTA >

DELTA section is size capped to a maximum of 32 (subject to change) and is written
from right to left, whereas MAIN entries are written from left to right.
Insertions which should normally end up in the MAIN section, but would be more
efficient to put in the DELTA section ends up there instead. When DELTA section
becomes full it's consolidated into the MAIN section using a merge join.
This co-operative merge is the core of the performance improvement.
Delta section doesn't occupy space if empty and if necessary, say in a completely
ordered insertion order, the main section can grow to fill the entire node.
This works automatically since each section writes from each direction in the node.

Seeker checks both sections when iterating over entries. To keep down complexity
nodes are always consolidated before doing merge, rebalance or split.

Removals cannot yet end up as removals in the delta section, so worst case for
removals hasn't improved in this change set. That is for the future.