Directly moves write cursor the last move onto leaf
Instead of always moving read cursor down and then having write cursor follow.
This reduces the number of pins. This works by remembering the depth of the tree
after the first write. This could be enhanced further by keeping depth in
tree state such that even the first write for a writer would take advantage of this.