Reads correct key when merging leaf nodes
That method is currently only called when underflowing and the right sibling
has exactly max/2 keyCount. In this scenario the rightSinblingKeyCount - 1 was
always correct, but in a scenario where multiple keys would be removed in one
go that pos was actually incorrect.

This is now corrected by reading the left-most key in the right sibling before
the merge. The behaviour hasn't changed, but the lurking bug which would
perhaps be exposed at a later time in another changeset is fixed.