WIP Move storage engine API to kernel-api
Ideally the storage engine API would live in its own component, but as things
currently are there's an overlap between storage engine API and kernel API and
so they are better located together. This move is very raw and imperfect and
is merely a start of finding a good cut-off point, isolating API from impl in
a reasonably sane way.

The immediate benefit of having API and impl split like this is that no storage
details specific to a particular storage engine can sneak/leak into the API
without the compiler noticing. This cut-off point is the first step in a bigger
picture where the record storage engine should also leave kernel and with that
things like importer, which will from that point on be specific to the
record storage engine, of course with the possibility to extract common or
generic features out to lib components.