WIP A start of shaping up everything regarding the entry header
Methods to attack and properly make care for entry sizes
that goes beyond the max TreeNode entry size are:
- DynamicSizeUtil#putKeyValueHeader
- DynamicSizeUtil#readKeyValueHeader
- DynamicSizeUtil#getEntryOverhead

and start testing from DynamicSizeUtilTest,
then TreeNodeDynamicSizeTest and then upwards.

There's still a "problem" with TreeNode#valueAt and that there
are paths leading to it which aren't easy to change. It's possible
but may involve some bigger changes. It's OK though since getting
rid of naked calls to it can be considered an optimization.