Simplifies token holders
By using the new NamedToken everywhere. The old Token is removed
as well as its custom subclasses. Specific types of token instances
were used as generics in lots of places throughout the whole stack
for the sole purpose of keeping RelationshipTypeToken instances
that also implemented RelationshipType so that they could be handed
out via the embedded API w/o creating additional instances.
This wasn't done for labels btw and we so not worth it when considering
all the overhead and specialization in so many places.

The token holders are no longer resolvable via the DependencyResolver either.