Removes token related methods from StorageReader
And at the same time breaks the dependency between IdGeneratorFactory
and TokenCreator. This after a realization that TokenHolder
is a generic kernel component. The only thing it needs from storage
is a way to generate new token ids, which StorageReader now provides,
just as it have done for nodes/relationship ids for a long time.